% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-meta.R
\name{detect_sensitivity}
\alias{detect_sensitivity}
\title{Detect single-study influence}
\usage{
detect_sensitivity(x)
}
\arguments{
\item{x}{A \code{meta} object.}
}
\description{
Detect whether meta-analysis is being influenced by a single study. If so, remove the study from the overall results.
}
\keyword{internal}
