% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic-package.R
\docType{package}
\name{metabolic-package}
\alias{metabolic}
\alias{metabolic-package}
\title{metabolic: Datasets and Functions for Reproducing Meta-Analyses}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Dataset and functions from the meta-analysis published in Medicine & Science in Sports & Exercise. It contains all the data and functions to reproduce the analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fmmattioni/metabolic}
  \item Report bugs at \url{https://github.com/fmmattioni/metabolic/issues}
}

}
\author{
\strong{Maintainer}: Felipe Mattioni Maturana \email{felipe.mattioni@med.uni-tuebingen.de} (\href{https://orcid.org/0000-0002-4221-6104}{ORCID}) [copyright holder]

}
\keyword{internal}
