% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{perform_bind}
\alias{perform_bind}
\title{Combine the subgroup meta-analyses}
\usage{
perform_bind(x)
}
\arguments{
\item{x}{An object retrieved from \link[=perform_meta]{perform_meta}.}
}
\value{
a \link[tibble:tibble-package]{tibble} with named lists.
}
\description{
Combine the subgroup meta-analyses to ...
}
\examples{
if (interactive()) {
# Perform meta-analysis on VO2max
results <- perform_meta(endpoint = "VO2max")
results

# Combine Overall and Subgroups meta-analysis results
results_bind <- perform_bind(results$meta_analysis)
results_bind
}
}
