% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-paper.R
\name{read_paper}
\alias{read_paper}
\title{Read the paper}
\usage{
read_paper()
}
\description{
This function will open the published paper in the journal website for you to read it in your default browser.
}
\examples{
read_paper()

}
